+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * IntuneDeviceConfigurationCustomPolicyWindows10
   * Add support to decrypt encrypted OmaSettings and export them in plaintext
     FIXES [#3655](https://github.com/microsoft/Microsoft365DSC/issues/3655)
+* IntuneDeviceEnrollmentPlatformRestriction
+  * Fix Set-TargetResource due to an issue were the bodyparameter not cast correctly
+  FIXES [#3730](https://github.com/microsoft/Microsoft365DSC/issues/3730)
 * IntuneEndpointDetectionAndResponsePolicyWindows10
   * Fix issue when trying to remove policy and Identity is set to a random GUID
     or from another tenant